requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Database connectivity
sqlalchemy>=2.0.0

# ER diagram generation (Optional - causes issues on Windows)
# eralchemy2>=1.3.0

# Graph analysis
networkx>=3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Table formatting
tabulate>=0.9.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests (for OpenRouter API)
requests>=2.31.0

# Data Quality Overhaul
duckdb>=0.9.0
plotly>=5.15.0
streamlit>=1.25.0
